requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.12.0
python-multipart==0.0.6
moviepy==1.0.3 
httpx==0.27.2
pydub==0.25.1
celery==5.3.6
redis==5.0.1
flask-session==0.5.0